refactor: move paths to util
